@@ -4,8 +4,10 @@
   "private": true,
   "dependencies": {
     "emailjs-com": "^3.2.0",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "framer-motion": "^10.16.4",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-icons": "^4.11.0",
     "react-scripts": "5.0.1"
   },
   "devDependencies": {
