@@ -1,25 +1,57 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import resumeData from '../data/resumeData';
+import { motion, AnimatePresence } from 'framer-motion';
+import { FaBars, FaTimes } from 'react-icons/fa';
 import './NavBar.css';
 
 /**
- * Primary navigation bar. Anchors link to different sections in the page.
- * Feel free to extend or replace this component with your own styling or
- * navigation logic. The links correspond to ids defined in App.js.
+ * Modern Navigation Bar with Glassmorphism and Mobile Responsive Menu
  */
 function NavBar() {
   const { name } = resumeData.personal;
+  const [isOpen, setIsOpen] = useState(false);
+  const [scrolled, setScrolled] = useState(false);
+
+  // Handle scroll effect for navbar
+  useEffect(() => {
+    const handleScroll = () => {
+      setScrolled(window.scrollY > 50);
+    };
+    window.addEventListener('scroll', handleScroll);
+    return () => window.removeEventListener('scroll', handleScroll);
+  }, []);
+
+  const toggleMenu = () => setIsOpen(!isOpen);
+
+  const navLinks = [
+    { title: 'Home', href: '#home' },
+    { title: 'About', href: '#about' },
+    { title: 'Skills', href: '#skills' },
+    { title: 'Experience', href: '#experience' },
+    { title: 'Contact', href: '#contact' }
+  ];
+
   return (
-    <nav className="navbar">
-      <div className="navbar-brand">{name}</div>
-      <ul className="navbar-links">
-        <li><a href="#home">Home</a></li>
-        <li><a href="#about">About</a></li>
-        <li><a href="#skills">Skills</a></li>
-        <li><a href="#experience">Experience</a></li>
-        {/*<li><a href="#projects">Projects</a></li>*/}
-        <li><a href="#contact">Contact</a></li>
-      </ul>
+    <nav className={`navbar ${scrolled ? 'scrolled' : ''}`}>
+      <div className="navbar-container">
+        <a href="#home" className="navbar-logo">
+          {name.split(' ')[0]}<span>.</span>
+        </a>
+
+        <div className="menu-icon" onClick={toggleMenu}>
+          {isOpen ? <FaTimes /> : <FaBars />}
+        </div>
+
+        <ul className={isOpen ? 'nav-menu active' : 'nav-menu'}>
+          {navLinks.map((link, index) => (
+            <li key={index} className="nav-item">
+              <a href={link.href} className="nav-links" onClick={() => setIsOpen(false)}>
+                {link.title}
+              </a>
+            </li>
+          ))}
+        </ul>
+      </div>
     </nav>
   );
 }
